Separate webpack config for library build
So we can build with NODE_ENV=development without hot reloading
